+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Phony changelog
 
+## Next release
+
+- **[IMPROVED]** Refactored the feature detector.
+
 ## 0.14.4 (2016-11-23)
 
 - **[FIXED]** Fixed mocking of classes with `self` return type ([#208]).